Update to Godot 4.4.1. Rework tile data collection.
Updating the project to Godot 4.4.1. Tilemap is no longer used so changing the test 1 into tilemap layers under a Node 2D. Rewrote the code for gathering tiles data at the beginning of the levels, remove hidden tiles. Tiles are now stored in a dictionary with its coordinates as keys (z=0, x=1, y =2). Rewrote the debugging printout for better clarity.