Rework tile collection. Add layer change feature.
Reworked the functions for collecting all tiles in use, all visible tiles, and hidden tiles. Created corresponding dictionaries which will be used at the beginning of the level.

Add functions to allow for players to move between environmental layers. Dimming higher layers for the layers below to be visible. If players are at the highest layer then all layers are visible.